requirements.txt
Django==3.0.2
djangorestframework==3.11.0
django-cors-headers==3.2.1
drf-spectacular==0.27.2
python-decouple==3.4